Add files via upload
Similar to EOMs, except each OEM appears in a row, and each year since 1903 appears in a column.  
For a given i,j, i is the row of a given OEM, and the jth column corresponds to a given year. The value of i,j is either 0 or 1: if the OEM did not exist in that year, it is 0. If the OEM did exist in that year, it is 1